requirements.txt
aniso8601==9.0.1
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
flake8==3.9.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-JWT==0.3.2
Flask-JWT-Extended==4.1.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
jwt==1.2.0
MarkupSafe==1.1.1
mccabe==0.6.1
passlib==1.7.4
pkg_resources==0.0.0
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.0.1
python-dotenv==0.17.0
pytz==2021.1
six==1.15.0
SQLAlchemy==1.3.24
sqlalchemy-filters==0.12.0
SQLAlchemy-Utils==0.37.0
Werkzeug==1.0.1
